createFileObj created for outsourcing the file object creatation File name duplicates are not allowed anymore
