+++ public/src/components/AddFileModalContent.js
@@ -3,11 +3,10 @@ import Fa from './Fa'
 import { faPlus  } from '@fortawesome/free-solid-svg-icons'
 import '../styles/AddFileModalContent.css'
 export default function AddFileModalContent(props){
-    let [fileName , setFileName] = useState('');
     return(
         <div className="AddFileModalContainer">
             <div className="AddFileModalInputs">
-                <input value={fileName} onChange={e => setFileName(e.target.value)} type="text" className="AddFileModalInput" id="AddFileModalInput" /> 
+                <input type="text" className="AddFileModalInput" id="AddFileModalInput" /> 
                 <button className="AddFileModalButton" onClick={e => props.addToFiles(document.querySelector('#AddFileModalInput').value)}><Fa icon={faPlus} /></button>
             </div>
             {props.validationError && (
