+++ public/src/helpers/isDuplicate.js
@@ -1,0 +1,4 @@
+export default (name, extention , listOfFileNames) => {
+    const fullFileName = name + '.' + extention
+     return listOfFileNames.filter(item => item === fullFileName).length > 0 ? true : false
+}
